refactor: root component should be hmr friendly
