+++ packages/core/src/runtime/clientEntry.tsx
@@ -1,9 +1,6 @@
-import { BrowserRouter, DataContext, ThemeContext } from '@rspress/runtime';
 import { isProduction } from '@rspress/shared';
-import { useThemeState } from '@theme';
-import { useMemo, useState } from 'react';
 import siteData from 'virtual-site-data';
-import { App, initPageData } from './App';
+import { RootApp } from './ClientApp';
 
 const enableSSG = siteData.ssg;
 
@@ -12,28 +9,6 @@ const enableSSG = siteData.ssg;
 export async function renderInBrowser() {
   const container = document.getElementById('root')!;
 
-  const enhancedApp = async () => {
-    const initialPageData = await initPageData(window.location.pathname);
-
-    return function RootApp() {
-      const [data, setData] = useState(initialPageData);
-      const [theme, setTheme] = useThemeState();
-      return (
-        <ThemeContext.Provider
-          value={useMemo(() => ({ theme, setTheme }), [theme, setTheme])}
-        >
-          <DataContext.Provider
-            value={useMemo(() => ({ data, setData }), [data, setData])}
-          >
-            <BrowserRouter>
-              <App />
-            </BrowserRouter>
-          </DataContext.Provider>
-        </ThemeContext.Provider>
-      );
-    };
-  };
-  const RootApp = await enhancedApp();
   if (process.env.__REACT_GTE_18__) {
     const { createRoot, hydrateRoot } = require('react-dom/client');
     if (isProduction() && enableSSG) {
